Some updates to “closest point in direction”
The idea is simple: We need yet another data structure to store all of
the potential indexed points as some type of grid that doesn’t
constantly reallocate. Yes, we need to be a bit efficient here, can
already see slowdowns with only 2 blobs. Really, none of the data
structures should reallocate. Unnecessary textured drawing and
“modulate color” are definitely the bottleneck though.